requirements-development.txt
-e .
pytest>=2.9.0,<3.0
pytest-django
pytest-factoryboy
Faker
recommonmark
Sphinx
sphinx_rtd_theme
tox
mock
django-debug-toolbar
